[AIP-21] Move operators.redshift_to_s3_operator providers.amazon.aws.operators.redshift_to_s3
